Split Survey variant into separate component
