+++ src/components/MessageCard/MessageCard.Survey.variants.js
@@ -1,0 +1,3 @@
+import { MessageCardSurveyThumbs as ThumbsSurvey } from './components/MessageCard.Survey.Thumbs'
+
+export { ThumbsSurvey }
