+++ src/components/MessageCard/components/MessageCard.Survey.jsx
@@ -1,18 +1,18 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 import {
   ConfirmationMessageUI,
-  EmojiButtonUI,
   FeedbackFormUI,
   SpinnerContainerUI,
-  SurveyOptionsUI,
   SurveyUI,
-} from '../MessageCard.css'
+} from './MessageCard.Survey.css'
 import Input from '../../Input'
 import Button from '../../Button'
 import Spinner from '../../Spinner'
 import Icon from '../../Icon'
 
 export const MessageCardSurvey = ({
+  children,
   withFeedbackForm = false,
   forceFeedbackForm = false,
   feedbackFormText = '',
@@ -20,17 +20,12 @@ export const MessageCardSurvey = ({
   showSpinner = false,
   showConfirmationMessage = false,
 }) => {
-  const buttons = [
-    { id: 'yes', text: '👍' },
-    { id: 'no', text: '👎' },
-  ]
-
-  const [showFeedbackForm, setShowFeedbackForm] = React.useState(false)
-  const [selected, setSelected] = React.useState(null)
   const [feedback, setFeedback] = React.useState('')
+  const [selected, setSelected] = React.useState(null)
+  const [showFeedbackForm, setShowFeedbackForm] = React.useState(false)
   const shouldShowFeedbackForm = showFeedbackForm || forceFeedbackForm
 
-  function handleClick(id) {
+  function handleSelection(id) {
     setSelected(id)
 
     if (withFeedbackForm) {
@@ -51,6 +46,15 @@ export const MessageCardSurvey = ({
     })
   }
 
+  function renderChildren() {
+    return React.Children.map(children, child => {
+      return React.cloneElement(child, {
+        selected,
+        onSelection: handleSelection,
+      })
+    })
+  }
+
   if (showConfirmationMessage) {
     return (
       <SurveyUI data-cy="beacon-message-cta-survey">
@@ -64,19 +68,7 @@ export const MessageCardSurvey = ({
 
   return (
     <SurveyUI data-cy="beacon-message-cta-survey">
-      <SurveyOptionsUI>
-        {buttons.map(({ id, text }) => (
-          <EmojiButtonUI
-            key={id}
-            onClick={() => handleClick(id)}
-            selected={selected === id}
-          >
-            <span role="img" aria-label="cool">
-              {text}
-            </span>
-          </EmojiButtonUI>
-        ))}
-      </SurveyOptionsUI>
+      {renderChildren()}
 
       {shouldShowFeedbackForm && (
         <FeedbackFormUI onSubmit={handleSubmit}>
@@ -84,9 +76,9 @@ export const MessageCardSurvey = ({
           <Input
             id="survey-comment"
             name="survey-comment"
-            multiline={3}
             withCharValidator={true}
             charValidatorLimit={120}
+            multiline={3}
             value={feedback}
             onChange={value => setFeedback(value)}
           />
@@ -103,4 +95,14 @@ export const MessageCardSurvey = ({
       )}
     </SurveyUI>
   )
+}
+
+MessageCardSurvey.propTypes = {
+  children: PropTypes.node.isRequired,
+  withFeedbackForm: PropTypes.bool,
+  forceFeedbackForm: PropTypes.bool,
+  feedbackFormText: PropTypes.string,
+  onSubmit: PropTypes.func,
+  showSpinner: PropTypes.bool,
+  showConfirmationMessage: PropTypes.bool,
 }
